{{cookiecutter.project_slug}}/prediction_engine/requirements.txt
aioredis
python-dotenv
sqlalchemy
scikit-learn
pandas
keras
xgboost
pyarrow